apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mixedworld-api"
version = "0.1.0"
description = "FastAPI backend for the MixedWorld MVP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.30,<1.0",
  "pydantic>=2.9,<3.0",
  "python-dotenv>=1.0,<2.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
